refactor(gsd): extract auto-worktree conflict policy
